Move functionality folder out of Web into main
